Added an initial set-up for Order Management
Added a "viewOrder" page which dispalys any items added to the customer's cart and allows the customer to checkout the order. Updated all other main pages with a "View Cart [#]" button to take them to the viewOrder page. Updated CSS for the viewOrder page. Added a HashMap creation for the cart once the user logs in or registers to track the cart throughout the website.